Evolve directory structure
This repository is taking a monorepo shape and it's necessary to proceed with another evolution of the directory structure. The idea is to have all the deployable applications under `apps` and rest of the NPM packages under `packages`. I have this in my head ever since we did the last evolution but we didn't have more apps in the monorepo so it didn't make much sense. But now we have it so it's the right time to do it.

It may seem a little bit odd to have src/apps inside yet another src/apps. But I believe this is fine. It's because root `src` contains our applications and packages separately. All these workspaces are actually repositories as well so it makes sense to start with `src` again just like we already do with packages. Good thing is that this uncovered some inconsistencies in our dependencies (because we didn't have everything from GraphQL together).

kiwicom-source-id: 791e6f862c2d2b32ffac00e6963d2cc49c90fa56