Tests: do not block stdin in watch mode
Now it's possible to interact with the watch mode because I remove the pipe input. The script is now calling Yarn Workspaces on it's own. I also removed the additional reporters because they were messing with the Jest output (we tried, not so useful - we have better eslint now and I know what tests are slow even without it). There is still a lot to improve but I like how it's getting better and better shape every time... :)

kiwicom-source-id: 9217a4115fecb3c5edf9fb127810b27681b5e67b