Fix tests runner execution
I noticed it didn't work because I applied the external config to naively. So anytime when someone use for example `--ci` it just didn't work. It's fixed now so the algorithm works again. I also extracted the inner logic to separate component just to leave the code in a better shape.

kiwicom-source-id: 786f11a7fdcbeb0a61e07ff3399e3b3da4210883