Monorepo utils: improve detection of changed files
It is sometimes needed to build a branch on an existing branch to be able to:
- use the code already prepared for merge
- prevent too many or too complicated rebase/merge conflicts

This commit improves the detection of the branch the current branch is built on (for most use cases it will detect master as it did before). However, for some use cases it allows testing/linting files that are changed in the branch itself, not all files changed in the previous branches it is built on. It is very useful when building a branch on another branch where a file which triggers testing all files is changed (e.g. package.json, .eslintrc).

The change expects that the developer uses rebases instead of merging to use changes from master branch in the feature branch.

kiwicom-source-id: 799bbc60f4d1cd120ae6899a5776b48d468d8b3c